Changed File Names
Files names were updated to Spencer's ideas
