+++ src/main/java/frc/robot/RobotContainer.java
@@ -25,8 +25,8 @@ import edu.wpi.first.wpilibj.DataLogManager;
  */
 public class RobotContainer {
   private Joystick controller = new Joystick(Constants.JoystickPort);
-  private Limelight limelight = new Limelight();
-  private Drivetrain drivetrain = new Drivetrain();
+  private SUB_Limelight limelight = new SUB_Limelight();
+  private SUB_Drivetrain drivetrain = new SUB_Drivetrain();
 
   
   JoystickButton C_aButton = new JoystickButton(controller, 1);
@@ -57,7 +57,7 @@ public class RobotContainer {
     
     configureBindings();
     limelight.setLed(1);
-    C_aButton.whileTrue(new LimeSequential(drivetrain, limelight));
+    C_aButton.whileTrue(new CMD_LimeSequential(drivetrain, limelight));
   }
 
   /**
@@ -74,7 +74,7 @@ public class RobotContainer {
 
     new Trigger(m_exampleSubsystem::exampleCondition)
         .onTrue(new ExampleCommand(m_exampleSubsystem));
-        drivetrain.setDefaultCommand(new TeleDrive(drivetrain, () -> -controller.getRawAxis(1),
+        drivetrain.setDefaultCommand(new CMD_TeleDrive(drivetrain, () -> -controller.getRawAxis(1),
                                 () -> -controller.getRawAxis(4)));
     // Schedule `ExampleCommand` when `exampleCondition` changes to `true`
 
@@ -99,6 +99,6 @@ public class RobotContainer {
    */
   public Command getAutonomousCommand() {
     // An example command will be run in autonomous
-    return new LimeSequential(drivetrain, limelight);
+    return new CMD_LimeSequential(drivetrain, limelight);
   }
 }
