+++ src/main/java/frc/robot/commands/CMD_TeleDrive.java
@@ -4,13 +4,13 @@ import java.util.function.Supplier;
 
 import com.revrobotics.CANSparkMax.IdleMode;
 
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.SUB_Drivetrain;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class TeleDrive extends CommandBase {
+public class CMD_TeleDrive extends CommandBase {
   @SuppressWarnings({ "PMD.UnusedPrivateField", "PMD.SingularField" })
-  private Drivetrain drive;
+  private SUB_Drivetrain drive;
   private Supplier<Double> Left, Right;
 
   /**
@@ -18,7 +18,7 @@ public class TeleDrive extends CommandBase {
    *
    * @param subsystem The subsystem used by this command.
    */
-  public TeleDrive(Drivetrain drivetrain, Supplier<Double> L, Supplier<Double> R) {
+  public CMD_TeleDrive(SUB_Drivetrain drivetrain, Supplier<Double> L, Supplier<Double> R) {
     this.drive = drivetrain;
     this.Left = L;
     this.Right = R;
